crawler-service/requirements.txt
scrapy==2.11.0
kafka-python==2.0.2
redis==4.6.0